requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
python-multipart
PyJWT
passlib[bcrypt]
python-jose
cryptography
